+++ core/src/main/resources/logback.xml
@@ -41,7 +41,7 @@
         <appender-ref ref="console"/>
         <appender-ref ref="SYSLOG"/>
     </logger>
-    <logger name="org.broadinstitute.dsde.rawls.dataaccess.HttpGooglePubSubDAO" level="debug" additivity="false">
+    <logger name="org.broadinstitute.dsde.rawls.google.HttpGooglePubSubDAO" level="debug" additivity="false">
         <appender-ref ref="file"/>
         <appender-ref ref="console"/>
         <appender-ref ref="SYSLOG"/>
