+++ core/src/main/scala/org/broadinstitute/dsde/rawls/monitor/CreatingBillingProjectMonitor.scala
@@ -5,7 +5,7 @@ import akka.actor.{Actor, Props}
 import akka.pattern._
 import com.typesafe.scalalogging.LazyLogging
 import org.broadinstitute.dsde.rawls.dataaccess.slick.{RawlsBillingProjectOperationRecord, RawlsBillingProjectRecord}
-import org.broadinstitute.dsde.rawls.dataaccess.{GooglePubSubDAO, ProjectTemplate, GoogleServicesDAO, SlickDataSource}
+import org.broadinstitute.dsde.rawls.dataaccess.{ProjectTemplate, GoogleServicesDAO, SlickDataSource}
 import org.broadinstitute.dsde.rawls.model._
 import org.broadinstitute.dsde.rawls.monitor.CreatingBillingProjectMonitor.{CheckDone, CheckNow}
 import scala.concurrent.duration._
