+++ google/src/main/scala/org/broadinstitute/dsde/rawls/google/HttpGooglePubSubDAO.scala
@@ -1,14 +1,14 @@
-package org.broadinstitute.dsde.rawls.dataaccess
+package org.broadinstitute.dsde.rawls.google
 
 import akka.actor.ActorSystem
 import com.google.api.client.auth.oauth2.Credential
-import com.google.api.client.googleapis.auth.oauth2.{GoogleCredential, GoogleClientSecrets}
+import com.google.api.client.googleapis.auth.oauth2.{GoogleClientSecrets, GoogleCredential}
 import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport
 import com.google.api.client.http.HttpResponseException
 import com.google.api.client.json.jackson2.JacksonFactory
-import com.google.api.services.pubsub.{Pubsub, PubsubScopes}
 import com.google.api.services.pubsub.model._
-import org.broadinstitute.dsde.rawls.dataaccess.GooglePubSubDAO._
+import com.google.api.services.pubsub.{Pubsub, PubsubScopes}
+import org.broadinstitute.dsde.rawls.google.GooglePubSubDAO._
 import org.broadinstitute.dsde.rawls.util.FutureSupport
 import spray.http.StatusCodes
